Add new style handling to histogram fix #457
